+++ glassfish-managed-4/src/main/java/org/jboss/arquillian/container/glassfish/managed_4/GlassFishManagedDeployableContainer.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.arquillian.container.glassfish.managed_4_1;
+package org.jboss.arquillian.container.glassfish.managed_4;
 
 import java.io.BufferedReader;
 import java.io.IOException;
