+++ glassfish-managed-4/src/main/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,1 +1,1 @@
-org.jboss.arquillian.container.glassfish.managed_4_1.GlassFishManagedContainerExtension
+org.jboss.arquillian.container.glassfish.managed_4.GlassFishManagedContainerExtension
