+++ glassfish-managed-4/src/test/java/org/jboss/arquillian/container/glassfish/managed_4/CDIJarTestCase.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.arquillian.container.glassfish.managed_4_1;
+package org.jboss.arquillian.container.glassfish.managed_4;
 
 import javax.inject.Inject;
